Remove AbstractArrayList
